requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioresponses==0.7.8
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kombu==5.5.3
motor==3.7.0
multidict==6.4.3
packaging==25.0
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pymongo==4.12.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.3.0
respx==0.22.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
